# New `bf16` & faster `i8` kernels ❤️ AMD Genoa