--- a/.vscode/launch.json
+++ b/.vscode/launch.json
@@ -11,7 +11,7 @@
             "program": "${workspaceFolder}/build_debug/simsimd_bench",
             "cwd": "${workspaceFolder}",
             "args": [
-                "--benchmark_filter=i8_cos"
+                "--benchmark_filter=cos_i8"
             ],
             "environment": [
                 {