--- a/binding.gyp
+++ b/binding.gyp
@@ -1,13 +1,11 @@
 {
-    "variables": {
-        "openssl_fips": ""
-    },
+    "variables": {"openssl_fips": ""},
     "targets": [
         {
             "target_name": "simsimd",
             "sources": ["javascript/lib.c"],
             "include_dirs": ["include"],
-            "defines": ["SIMSIMD_NATIVE_F16=0"],
+            "defines": ["SIMSIMD_NATIVE_F16=0", "SIMSIMD_NATIVE_BF16=0"],
             "cflags": [
                 "-std=c11",
                 "-ffast-math",
@@ -17,20 +15,17 @@
                 "-Wno-switch",
             ],
             "conditions": [
-                ["OS=='mac'", {
-                    "xcode_settings": {
-                        "MACOSX_DEPLOYMENT_TARGET": "11.0",
-                        "OTHER_CFLAGS": [
-                            "-arch arm64",
-                            "-arch x86_64"
-                        ],
-                        "OTHER_LDFLAGS": [
-                            "-arch arm64",
-                            "-arch x86_64"
-                        ]
-                    }
-                }]
-            ]
+                [
+                    "OS=='mac'",
+                    {
+                        "xcode_settings": {
+                            "MACOSX_DEPLOYMENT_TARGET": "11.0",
+                            "OTHER_CFLAGS": ["-arch arm64", "-arch x86_64"],
+                            "OTHER_LDFLAGS": ["-arch arm64", "-arch x86_64"],
+                        }
+                    },
+                ]
+            ],
         }
-    ]
+    ],
 }