--- a/c/lib.c
+++ b/c/lib.c
@@ -5,6 +5,7 @@
  */
 #define SIMSIMD_DYNAMIC_DISPATCH 1
 #define SIMSIMD_NATIVE_F16 0
+#define SIMSIMD_NATIVE_BF16 0
 
 /*  Depending on the Operating System, the following intrinsics are available
  *  on recent compiler toolchains: