--- a/cpp/test.c
+++ b/cpp/test.c
@@ -8,6 +8,7 @@
 #include <stdio.h>  // `printf`
 
 #define SIMSIMD_NATIVE_F16 0
+#define SIMSIMD_NATIVE_BF16 0
 #define SIMSIMD_RSQRT(x) (1 / sqrtf(x))
 #define SIMSIMD_LOG(x) (logf(x))
 #include <simsimd/simsimd.h>
@@ -22,6 +23,7 @@ void print_capabilities(void) {
     char const* flags[2] = {"false", "true"};
     printf("Benchmarking Similarity Measures\n");
     printf("- Compiler used native F16: %s\n", flags[SIMSIMD_NATIVE_F16]);
+    printf("- Compiler used native BF16: %s\n", flags[SIMSIMD_NATIVE_BF16]);
     printf("\n");
     printf("Compile-time settings:\n");
     printf("- Arm NEON support enabled: %s\n", flags[SIMSIMD_TARGET_NEON]);
@@ -55,6 +57,7 @@ void test_utilities(void) {
     int uses_skylake = simsimd_uses_skylake();
     int uses_ice = simsimd_uses_ice();
     int uses_sapphire = simsimd_uses_sapphire();
+    int uses_genoa = simsimd_uses_genoa();
 
     assert(uses_neon == ((capabilities & simsimd_cap_neon_k) != 0));
     assert(uses_sve == ((capabilities & simsimd_cap_sve_k) != 0));
@@ -62,6 +65,7 @@ void test_utilities(void) {
     assert(uses_skylake == ((capabilities & simsimd_cap_skylake_k) != 0));
     assert(uses_ice == ((capabilities & simsimd_cap_ice_k) != 0));
     assert(uses_sapphire == ((capabilities & simsimd_cap_sapphire_k) != 0));
+    assert(uses_genoa == ((capabilities & simsimd_cap_genoa_k) != 0));
 }
 
 /**